@@ -4,12 +4,12 @@ plugins {
 
 android {
     namespace = "com.app.primeshow"
-    compileSdk = 33
+    compileSdk = 34
 
     defaultConfig {
         applicationId = "com.app.primeshow"
         minSdk = 30
-        targetSdk = 33
+        targetSdk = 34
         versionCode = 1
         versionName = "1.0"
 
