@@ -3,27 +3,29 @@ package com.app.primeshow.Activities;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
+import android.os.Build;
 import android.os.Bundle;
 import android.view.View;
+import android.view.WindowManager;
 
 import com.app.primeshow.databinding.ActivityIntroBinding;
 
 public class IntroActivity extends AppCompatActivity {
 
     ActivityIntroBinding binding;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         binding=ActivityIntroBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
-        binding.getin.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                startActivity(new Intent(IntroActivity.this, LoginActivity.class));
-                finish();
-            }
+        if(Build.VERSION.SDK_INT>=Build.VERSION_CODES.KITKAT){
+            getWindow().setFlags(WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS,WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS);
+        }
+
+        binding.getin.setOnClickListener(v -> {
+            startActivity(new Intent(IntroActivity.this, LoginActivity.class));
+            finish();
         });
     }
 }
